// src/models/types/context/auth.ts
import { User } from '../api/user';

export type AuthContextType = {
  isAdmin: boolean;
  setIsAdmin: (isAdmin: boolean) => void;
  isAuthenticated: boolean;
  accessToken: string | null;
  refreshToken: string | null;
  login: (accessToken: string, refreshToken: string) => void;
  logout: () => void;
  userData: User | null;
  setUserData: (userData: User | null) => void;
  getAuthentication: () => boolean;
  authenticationErrorCount: number;
};